start, move some types to mz_compute_types
* moves the ComputeReplicaConfig and ComputeReplicaLogging types to mz_compute_types crate
